SQLite support

+++ app/core/migrations/add_core_statistic_indexes.py
@@ -1,5 +1,8 @@
 from django.db import migrations
 
+from os import environ
+DJANGO_DEBUG = bool(environ.get('DJANGO_DEBUG', False))
+
 def add_indexes(apps, schema_editor):
     for x in range(0, 101):
         schema_editor.execute('CREATE INDEX core_statistic_value_%d ON core_statistic (core_statistic_compare(value::integer, %d))' % (x,x))
@@ -9,18 +12,20 @@ class Migration(migrations.Migration):
         ('core', '0001_initial')
     ]
 
-    operations = [
-    migrations.RunSQL('''
-        CREATE OR REPLACE FUNCTION core_statistic_compare(value integer, expectation integer) RETURNS double precision
-            LANGUAGE plpgsql IMMUTABLE
-            AS $$
-        DECLARE
-            accordance double precision;
-        BEGIN
-            RETURN (10 - ( abs(value - expectation) / 10));
-        END;
-        $$;
-    '''),
-    migrations.RunPython(add_indexes),
-    migrations.RunSQL('TRUNCATE TABLE core_statistic')
-    ]
+    if not DJANGO_DEBUG:
+
+        operations = [
+        migrations.RunSQL('''
+            CREATE OR REPLACE FUNCTION core_statistic_compare(value integer, expectation integer) RETURNS double precision
+                LANGUAGE plpgsql IMMUTABLE
+                AS $$
+            DECLARE
+                accordance double precision;
+            BEGIN
+                RETURN (10 - ( abs(value - expectation) / 10));
+            END;
+            $$;
+        '''),
+        migrations.RunPython(add_indexes),
+        migrations.RunSQL('TRUNCATE TABLE core_statistic')
+        ]
